Fix StatusNotifier bug after reload
Icons were not being removed after reloading the config as the match
rule to listen for removed services were being deleted.

This PR refactors the code and adds the ability to preserve match rules
when config is being reloaded.

Fixes #4811
