--- libqtile/resources/status_notifier/statusnotifieritem.py
@@ -1,51 +1,0 @@
-STATUS_NOTIFIER_ITEM_SPEC = """
-<!DOCTYPE node PUBLIC "-//freedesktop//DTD D-BUS Object Introspection 1.0//EN"
-"http://www.freedesktop.org/standards/dbus/1.0/introspect.dtd">
-<node>
-  <interface name='org.kde.StatusNotifierItem'>
-    <annotation name="org.gtk.GDBus.C.Name" value="Item" />
-    <method name='ContextMenu'>
-      <arg type='i' direction='in' name='x'/>
-      <arg type='i' direction='in' name='y'/>
-    </method>
-    <method name='Activate'>
-      <arg type='i' direction='in' name='x'/>
-      <arg type='i' direction='in' name='y'/>
-    </method>
-    <method name='SecondaryActivate'>
-      <arg type='i' direction='in' name='x'/>
-      <arg type='i' direction='in' name='y'/>
-    </method>
-    <method name='Scroll'>
-      <arg type='i' direction='in' name='delta'/>
-      <arg type='s' direction='in' name='orientation'/>
-    </method>
-    <signal name='NewTitle'/>
-    <signal name='NewIcon'/>
-    <signal name='NewAttentionIcon'/>
-    <signal name='NewOverlayIcon'/>
-    <signal name='NewToolTip'/>
-    <signal name='NewStatus'>
-      <arg type='s' name='status'/>
-    </signal>
-    <property name='Category' type='s' access='read'/>
-    <property name='Id' type='s' access='read'/>
-    <property name='Title' type='s' access='read'/>
-    <property name='Status' type='s' access='read'/>
-    <!-- See discussion on pull #536
-    <property name='WindowId' type='u' access='read'/>
-    -->
-    <property name='IconThemePath' type='s' access='read'/>
-    <property name='IconName' type='s' access='read'/>
-    <property name='IconPixmap' type='a(iiay)' access='read'/>
-    <property name='OverlayIconName' type='s' access='read'/>
-    <property name='OverlayIconPixmap' type='a(iiay)' access='read'/>
-    <property name='AttentionIconName' type='s' access='read'/>
-    <property name='AttentionIconPixmap' type='a(iiay)' access='read'/>
-    <property name='AttentionMovieName' type='s' access='read'/>
-    <property name='ToolTip' type='(sa(iiay)ss)' access='read'/>
-    <property name='Menu' type='o' access='read'/>
-    <property name='ItemIsMenu' type='b' access='read'/>
-  </interface>
-</node>
-"""
