+++ libqtile/widget/helpers/status_notifier/fallback_icon.png
